set the finished regions
