+++ rgen/rgen-biome/src/biome/README.md
@@ -1,0 +1,30 @@
+# Biomes
+All regions require documentation explaining biomes and geographic type.  
+
+**frozen_region**  
+- frozen_region/ice_cap  
+
+**cold_region**  
+- cold_region/tundra
+- cold_region/sub_arctic
+
+**cool_region**  
+- cool_region/cold_swamp  
+- cool_region/cool_temperate  
+
+**temperate_region**  
+- temperate_region/warm_temperate  
+- temperate_region/wet_temperate  
+
+**warm_region**  
+- warm_region/hot_swamp  
+- warm_region/tropical  
+
+**dry_region**  
+- dry_region/mediterranean  
+- dry_region/dry_temperate  
+- dry_region/savanna  
+
+**hot_region**  
+- hot_region/hot_desert  
+- hot_region/bad_lands
