--- rgen/rgen-biome/src/biome/forests/mod.rs
@@ -1,5 +1,0 @@
-mod birch;
-mod woodland;
-
-pub use birch::*;
-pub use woodland::*;
