+++ rgen/rgen-biome/src/biome/temperate_region/mod.rs
@@ -1,0 +1,3 @@
+mod warm_temperate;
+
+pub use warm_temperate::*;
